Trace conference-modify instead of room creation
